+++ Packages/com.texelsaur.video/Editor/Inspectors/Component/URL Source/PlaylistQueueInspector.cs
@@ -9,6 +9,7 @@ namespace Texel
         protected SerializedProperty priorityAccessProperty;
         protected SerializedProperty deleteAccessProperty;
 
+        protected SerializedProperty enableSyncQuestUrlsProperty;
         protected SerializedProperty syncTrackTitlesProperty;
         protected SerializedProperty syncTrackAuthorsProperty;
         protected SerializedProperty syncPlayerNamesProperty;
@@ -20,6 +21,7 @@ namespace Texel
             priorityAccessProperty = serializedObject.FindProperty(nameof(PlaylistQueue.priorityAccess));
             deleteAccessProperty = serializedObject.FindProperty(nameof(PlaylistQueue.deleteAccess));
 
+            enableSyncQuestUrlsProperty = serializedObject.FindProperty(nameof(PlaylistQueue.enableSyncQuestUrls));
             syncTrackTitlesProperty = serializedObject.FindProperty(nameof(PlaylistQueue.syncTrackTitles));
             syncTrackAuthorsProperty = serializedObject.FindProperty(nameof(PlaylistQueue.syncTrackAuthors));
             syncPlayerNamesProperty = serializedObject.FindProperty(nameof(PlaylistQueue.syncPlayerNames));
@@ -38,6 +40,7 @@ namespace Texel
 
             EditorGUILayout.Space();
             EditorGUILayout.LabelField("Sync", EditorStyles.boldLabel);
+            EditorGUILayout.PropertyField(enableSyncQuestUrlsProperty, new GUIContent("Sync Quest URLs", "When enabled, supports tracking separate Quest URLs if they are added through the API."));
             EditorGUILayout.PropertyField(syncTrackTitlesProperty);
             EditorGUILayout.PropertyField(syncTrackAuthorsProperty);
             EditorGUILayout.PropertyField(syncPlayerNamesProperty);